Moved game maze to its own class (wraps the maze class)
