+++ src/Phaser/Game/character.js
@@ -1,0 +1,5 @@
+export default class Character {
+  constructor(graphics, maze) {
+    this.graphics = graphics;
+  }
+}
